# example/docs/openapi.yaml
...
components:
  schemas:
...
    AddCommentReqBody:
      properties:
        authorId:
          type: string
        body:
          type: string
      required:
      - authorId
      - body
      type: object
...
    AddCommentRespBody:
      $ref: '#/components/schemas/Comment'
...
    AddContactReqBody:
      properties:
        contact:
          $ref: '#/components/schemas/Contact'
      type: object
...
    AddContactRespBody:
      $ref: '#/components/schemas/OkResp'
...
    AddMemberReqBody:
      properties:
        userId:
          type: string
      required:
      - userId
      type: object
...
    AddMemberRespBody:
      $ref: '#/components/schemas/OkResp'
...
    AuthRequiredErr:
      description: Unauthorized error response (HTTP 401).
      type: object
...
    Cents:
      type: integer
    Comment:
      description: |-
        Comment is a NESTED type that itself NESTS another cross-package
        type (`author users.UserRef`). The cascade demonstrates that
        arbitrarily-deep cross-package references resolve cleanly:
        Task → Comment → users.UserRef.
      properties:
        author:
          $ref: '#/components/schemas/UserRef'
        body:
          type: string
        createdAt:
          type: string
        id:
          type: string
      required:
      - id
      - body
      - createdAt
      type: object
...
    CreateOrderReqBody:
      properties:
        currency:
          type: string
        customer:
          $ref: '#/components/schemas/Customer'
        items:
          items:
            $ref: '#/components/schemas/LineItem'
          type: array
        metadata:
          additionalProperties:
            type: string
          type: object
        notes:
          type: string
        tags:
          items:
            type: string
          type: array
      type: object
    CreateOrderRespBody:
      $ref: '#/components/schemas/Order'
...
    CreateTaskReqBody:
      properties:
        assigneeId:
          type: string
        comments:
          items:
            $ref: '#/components/schemas/Comment'
          type: array
        projectId:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        title:
          type: string
      required:
      - projectId
      - title
      type: object
    CreateTaskRespBody:
      $ref: '#/components/schemas/Task'
...
    CreateUserReqBody:
      properties:
        avatar:
          $ref: '#/components/schemas/Avatar'
        bio:
          type: string
        email:
          type: string
        name:
          type: string
      required:
      - email
      - name
      type: object
    CreateUserRespBody:
      $ref: '#/components/schemas/User'
    CurrencyCode:
      type: string
    Customer:
      description: |-
        Customer is the level-2 nested type. Pairs cross-package refs
        (`referrer users.UserRef?`) with file uploads (`profilePic`)
        and the `@nullable` forced-pointer pattern on `loyaltyPoints`.
      properties:
        billingAddress:
          $ref: '#/components/schemas/Address'
        channels:
          additionalProperties:
            items:
              $ref: '#/components/schemas/Tag'
            type: array
          description: |-
            MAP-WITH-NESTED-COLLECTION: `map<string, Tag[]>` — each
            value is an ARRAY of Tag scalar. Codegen emits a doubly-
            nested loop pulling the scalar's per-element validators
            through both layers:

              for _, val := range v.Channels {
                  for i := range val { /* @minLength + @maxLength + @pattern on val[i] */ }
              }

            Use case: bucketing tag-style labels per channel
            ("email" → ["welcome", "newsletter"], "sms" → [...]).
          type: object
        email:
          $ref: '#/components/schemas/Email'
        gridLabels:
          description: |-
            MULTI-DIMENSIONAL ARRAY: `Tag[][]` — slice of slices. v1
            grammar accepts repeated `[]` suffixes; codegen emits the
            matching doubly-nested for-loops per scalar decorator.

              for i0 := range v.GridLabels {
                  for i1 := range v.GridLabels[i0] {
                      /* Tag's @pattern on v.GridLabels[i0][i1] */
                  }
              }

            Use case: row-and-column labelling of an analytics grid.
          items:
            items:
              $ref: '#/components/schemas/Tag'
            type: array
          type: array
        id:
          description: String + format(uuid).
          type: string
        index:
          additionalProperties:
            additionalProperties:
              $ref: '#/components/schemas/SafeURL'
            type: object
          description: |-
            NESTED MAP: `map<string, map<Tag, shared.SafeURL>>` — the
            outer map's value is itself a map whose key is a scalar
            (Tag) and value is a cross-package scalar (SafeURL).
            Codegen recursively walks both inner sides: 4 layers of
            wrapping per scalar decorator.

              for _, val0 := range v.Index {                 // outer map values
                  for k1 := range val0 { /* Tag's @pattern on k1 */ }
                  for _, val1 := range val0 { /* SafeURL's @format(url) on val1 */ }
              }
          type: object
        loyaltyPoints:
          description: |-
            @nullable on a numeric field forces a pointer in Go so the
            wire can carry an explicit `null` (different from "0 points
            means a brand-new account"). Without @required, presence is
            not enforced — caller may send the key with null OR omit it.
          nullable: true
          type: integer
        name:
          description: String + length range.
          type: string
        phone:
          description: |-
            4-STATE PRESENCE: optional + nullable + pattern. Phone
            matches a permissive E.164-ish regex when present.
          nullable: true
          type: string
        primaryAddress:
          $ref: '#/components/schemas/Address'
        profilePic:
          description: |-
            Optional FILE upload. Demonstrates the file-validator pair
            (`@maxSize` + `@mimeTypes`) which the field-type-compat
            pass restricts to `file` typed fields.
          format: binary
          type: string
        referrer:
          $ref: '#/components/schemas/UserRef'
      required:
      - id
      - email
      - name
      type: object
...
    Email:
      format: email
      type: string
    EmailTakenErr:
      description: Conflict error response (HTTP 409).
      properties:
        email:
          type: string
        existingId:
          type: string
      type: object
...
    InvalidTokenErr:
      description: Unauthorized error response (HTTP 401).
      type: object
...
    LineItem:
      properties:
        discount:
          $ref: '#/components/schemas/Discount'
        note:
          description: |-
            String with default. The handler's pre-fill logic sets the
            empty string before JSON decode so a missing `note` key
            surfaces as `""` rather than the Go zero value (which IS
            "", but the default gives the user a place to override).
          type: string
        productId:
          type: string
        quantity:
          type: integer
        sku:
          $ref: '#/components/schemas/SkuCode'
        unitCents:
          $ref: '#/components/schemas/Cents'
      required:
      - productId
      - sku
      - quantity
      - unitCents
      type: object
...
    ListOrdersRespBody:
      $ref: '#/components/schemas/OrderList'
...
    ListSuspendedUsersRespBody:
      $ref: '#/components/schemas/UserList'
...
    ListTasksRespBody:
      $ref: '#/components/schemas/TaskList'
...
    ListUsersRespBody:
      $ref: '#/components/schemas/UserList'
...
    MaintenanceWindowErr:
      description: ServiceUnavailable error response (HTTP 503).
      properties:
        estimatedEndAt:
          type: string
        reason:
          $ref: '#/components/schemas/MaintenanceWindowReason'
      type: object
    MaintenanceWindowReason:
      enum:
      - ScheduledMaintenance
      - IncidentRecovery
      - CapacityExpansion
      type: string
    MalformedRequestErr:
      description: BadRequest error response (HTTP 400).
      properties:
        column:
          type: integer
        detail:
          type: string
        line:
          type: integer
      type: object
...
    OkResp:
      description: |-
        OkResp is the conventional response shape for write endpoints
        that don't return a domain entity. Reusing it across services
        keeps the wire vocabulary tight: clients learn ONE success
        envelope, not one per service.
      properties:
        ok:
          type: boolean
      type: object
    Order:
      deprecated: true
      description: |-
        Order is the deeply-nested aggregate root. Demonstrates EVERY
        presence-state and almost every validator on its 14 fields.

        `@deprecated` at the type level surfaces in the generated Go as
        `// Deprecated: ...` so `staticcheck` flags every read of the
        type — useful for migrating clients off a legacy aggregate while
        keeping the schema available during the transition window.

        Scalar types (OrderID, CurrencyCode, ...) are referenced by name
        — every field of that scalar type inherits the validators from
        the scalar's declaration, so the field-level decorator chain
        stays focused on PRESENCE rules (`@required`, `@nullable`).
      properties:
        auditedBy:
          $ref: '#/components/schemas/UserRef'
        cosignedBy:
          $ref: '#/components/schemas/UserRef'
        createdAt:
          description: |-
            Required ISO timestamp at the wrapper level. Top-level
            datetime is always required for audit trails.
          type: string
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        customer:
          $ref: '#/components/schemas/Customer'
        id:
          $ref: '#/components/schemas/OrderID'
        items:
          description: |-
            Required slice of nested LineItems with bound. @minItems
            forbids "empty order"; @maxItems caps explosion attacks.
          items:
            $ref: '#/components/schemas/LineItem'
          type: array
        metadata:
          additionalProperties:
            $ref: '#/components/schemas/SafeURL'
          description: |-
            MAP-OF-SCALAR — both the key and the value are scalars.
            Codegen emits two per-element loops:

              for k := range v.Metadata     { /* Tag's @minLength + @maxLength + @pattern on k   */ }
              for _, val := range v.Metadata { /* SafeURL's @format(url) + @maxLength on val      */ }

            So the wire shape `{"primary": "https://...", ...}` rejects
            both an upper-case key (fails Tag pattern) and a non-URL
            value (fails SafeURL format) at the validator boundary.
          type: object
        notes:
          description: |-
            Optional string with maxLength validator. When omitted the
            field stays absent on the wire (`omitempty`).
          type: string
        payment:
          $ref: '#/components/schemas/Payment'
        shippedAt:
          description: |-
            4-STATE PRESENCE: optional + nullable. The wire payload may:
              - omit `shippedAt` entirely         (nil pointer, no key)
              - send `"shippedAt": null`          (nil pointer)
              - send `"shippedAt": "2026-01-..."` (validated as datetime)
            Validators run only when the value is non-null.
          nullable: true
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        tags:
          description: |-
            ARRAY OF SCALAR — `tags Tag[]` inherits the Tag scalar's
            per-element validators (@minLength, @maxLength, @pattern)
            automatically. The field-level chain only carries the
            array-level rules (@minItems, @maxItems, @uniqueItems).
          items:
            $ref: '#/components/schemas/Tag'
          type: array
        totalCents:
          $ref: '#/components/schemas/Cents'
      required:
      - id
      - totalCents
      - status
      - createdAt
      type: object
    OrderID:
      type: string
    OrderList:
      description: |-
        OrderList is the ListOrders response — same envelope contract
        as UserList / ProjectList / TaskList.
      properties:
        cursor:
          type: string
        items:
          items:
            $ref: '#/components/schemas/Order'
          type: array
        total:
          type: integer
      type: object
...
    RateLimitedErr:
      description: TooManyRequests error response (HTTP 429).
      properties:
        bucket:
          type: string
        retry_after:
          type: integer
      type: object
...
    SafeURL:
      format: url
      type: string
    SkuCode:
      type: string
...
    Tag:
      type: string
    Task:
      description: |-
        Task is the marquee multi-nested type. It demonstrates:

          - Cross-package required nested ref (`project projects.ProjectRef`)
          - Cross-package OPTIONAL nested ref via `T?` (`assignee users.UserRef?`)
            — the wire field is omitted when no one is assigned, which
            codegen renders as `*users.UserRef` in the Go struct.
          - Slice of nested type whose own fields cross another package
            (`comments Comment[]`, where each Comment carries a UserRef).
          - Enum field with default validation via the generated UnmarshalJSON.

        Validators on the wrapper (`@required`, `@length`, `@maxItems`)
        fire at the handler boundary; the nested types' own validators
        can be invoked explicitly from logic when deep validation is
        required (v1 doesn't auto-recurse).
      properties:
        assignee:
          $ref: '#/components/schemas/UserRef'
        comments:
          description: |-
            Slice of nested struct. Empty array is fine; @maxItems caps
            pathological growth.
          items:
            $ref: '#/components/schemas/Comment'
          type: array
        id:
          type: string
        project:
          $ref: '#/components/schemas/ProjectRef'
        status:
          $ref: '#/components/schemas/TaskStatus'
        title:
          type: string
      required:
      - id
      - title
      type: object
    TaskAssignmentConflictErr:
      description: Conflict error response (HTTP 409).
      properties:
        currentAssignee:
          $ref: '#/components/schemas/UserRef'
        pendingStatus:
          $ref: '#/components/schemas/TaskStatus'
      type: object
    TaskList:
      description: |-
        TaskList is the ListTasks response — same envelope contract as
        UserList / ProjectList for consistent client pagination.
      properties:
        cursor:
          type: string
        items:
          items:
            $ref: '#/components/schemas/Task'
          type: array
        total:
          type: integer
      type: object
...
    TaskStatus:
      enum:
      - open
      - in_progress
      - done
      - archived
      type: string
...
    UserList:
      description: |-
        UserList is the ListUsers response. Each list endpoint in the
        project wraps its items in a per-service envelope that carries
        a cursor + total — these field names stay identical across
        services so a generic client can paginate without per-resource
        special-casing.
      properties:
        cursor:
          type: string
        items:
          items:
            $ref: '#/components/schemas/User'
          type: array
        total:
          type: integer
      type: object
    UserNotFoundErr:
      description: NotFound error response (HTTP 404).
      type: object
    UserRef:
      description: |-
        UserRef is the lightweight reference shape — id + display name —
        embedded by other packages whenever they need to *mention* a user
        without dragging in the heavy User payload (avatar, bio, etc.).

        Cross-package nesting through UserRef is the canonical pattern for
        avoiding bloated wire payloads while keeping responses self-
        contained (no client-side join lookup needed).
      properties:
        id:
          type: string
        name:
          type: string
      required:
      - id
      - name
      type: object
...
    ValidationFailedErr:
      description: UnprocessableEntity error response (HTTP 422).
      properties:
        fields:
          items:
            type: string
          type: array
        hint:
          type: string
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
    oauth2:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  title: CraftGo Showcase
  version: 1.0.0
openapi: 3.1.0
paths:
  /orders:
    get:
      description: List orders with cursor pagination.
      operationId: ListOrders
      parameters:
      - in: query
        name: cursor
        schema:
          $ref: '#/components/schemas/ListOrdersReqQuery/properties/cursor'
      - in: query
        name: limit
        schema:
          $ref: '#/components/schemas/ListOrdersReqQuery/properties/limit'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOrdersRespBody'
          description: OK
        default:
          description: ""
      tags:
      - orders
    post:
      description: Create a new order. The server fills id, status, totalCents, auditedBy,
        and createdAt; the client supplies customer + items + optional payment metadata.
      operationId: CreateOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderReqBody'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderRespBody'
          description: OK
        default:
          description: ""
      tags:
      - orders
...
  /projects/{projectId}/members:
    post:
      description: Add a member to a project. Returns shared.OkResp; idempotent if
        the user is already a member.
      operationId: AddMember
      parameters:
      - in: path
        name: projectId
        required: true
        schema:
          $ref: '#/components/schemas/AddMemberReqPath/properties/projectId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberReqBody'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddMemberRespBody'
          description: OK
        default:
          description: ""
      tags:
      - projects
  /tasks:
    get:
      description: List tasks scoped to a project (use the projectId query parameter).
      operationId: ListTasks
      parameters:
      - in: query
        name: cursor
        schema:
          $ref: '#/components/schemas/ListTasksReqQuery/properties/cursor'
      - in: query
        name: limit
        schema:
          $ref: '#/components/schemas/ListTasksReqQuery/properties/limit'
      - in: query
        name: projectId
        schema:
          $ref: '#/components/schemas/ListTasksReqQuery/properties/projectId'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTasksRespBody'
          description: OK
        default:
          description: ""
      tags:
      - tasks
    post:
      description: Create a task under a project. Surfaces a 409 when the assignee
        already owns the slot — body carries the existing assignee via the cross-package
        users.UserRef type.
      operationId: CreateTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskReqBody'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTaskRespBody'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRequiredErr'
          description: Unauthorized
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskAssignmentConflictErr'
          description: Conflict
        default:
          description: ""
      tags:
      - tasks
...
  /tasks/{taskId}/comments:
    post:
      description: Append a comment to a task. The author UserRef is constructed by
        logic from the supplied authorId.
      operationId: AddComment
      parameters:
      - in: path
        name: taskId
        required: true
        schema:
          $ref: '#/components/schemas/AddCommentReqPath/properties/taskId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCommentReqBody'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddCommentRespBody'
          description: OK
        default:
          description: ""
      tags:
      - tasks
  /users:
    get:
      description: List users with cursor pagination. The request shape mirrors shared.Pagination
        — see the comment in shared/shared.craftgo for why the showcase keeps a local
        copy.
      operationId: ListUsers
      parameters:
      - in: query
        name: cursor
        schema:
          $ref: '#/components/schemas/ListUsersReqQuery/properties/cursor'
      - in: query
        name: limit
        schema:
          $ref: '#/components/schemas/ListUsersReqQuery/properties/limit'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersRespBody'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidTokenErr'
          description: Unauthorized
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitedErr'
          description: TooManyRequests
        default:
          description: ""
      tags:
      - users
    post:
      description: Create a new user. Surfaces a 409 when the email collides with
        an existing account, plus the standard auth/validation envelope.
      operationId: CreateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserReqBody'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserRespBody'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MalformedRequestErr'
          description: BadRequest
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidTokenErr'
          description: Unauthorized
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailTakenErr'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailedErr'
          description: UnprocessableEntity
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitedErr'
          description: TooManyRequests
        default:
          description: ""
      tags:
      - users
...
  /users/{userId}/contacts:
    post:
      description: Attach a Contact to a user. Demonstrates a nested type whose own
        type-level validators (@requiresOneOf / @mutuallyExclusive) gate the wire
        shape.
      operationId: AddContact
      parameters:
      - in: path
        name: userId
        required: true
        schema:
          $ref: '#/components/schemas/AddContactReqPath/properties/userId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddContactReqBody'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddContactRespBody'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MalformedRequestErr'
          description: BadRequest
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidTokenErr'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserNotFoundErr'
          description: NotFound
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailedErr'
          description: UnprocessableEntity
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceWindowErr'
          description: ServiceUnavailable
        default:
          description: ""
      tags:
      - users
  /users/suspended:
    get:
      description: List currently-suspended users with cursor pagination.
      operationId: ListSuspendedUsers
      parameters:
      - in: query
        name: cursor
        schema:
          $ref: '#/components/schemas/ListSuspendedUsersReqQuery/properties/cursor'
      - in: query
        name: limit
        schema:
          $ref: '#/components/schemas/ListSuspendedUsersReqQuery/properties/limit'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSuspendedUsersRespBody'
          description: OK
        default:
          description: ""
      security:
      - oauth2: []
      tags:
      - users
...
servers:
- url: /api