# example/internal/handler/order-service/list-orders-handler.go
// Code generated by craftgo. DO NOT EDIT.

package orderservice

import (
	"encoding/json"
	"github.com/dropship-dev/craftgo/pkg/server"
	"net/http"
	"strconv"

	logic "github.com/dropship-dev/craftgo/example/internal/logic/order-service"
	types "github.com/dropship-dev/craftgo/example/internal/types/orders"
	"github.com/dropship-dev/craftgo/example/svccontext"
)

// ListOrdersHandler returns the http.HandlerFunc for the
// GET ListOrders endpoint.
func ListOrdersHandler(svcCtx *svccontext.ServiceContext) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req types.ListOrdersReq
		if _v := r.URL.Query().Get("cursor"); _v != "" {
			req.Cursor = &_v
		}
		if _v := r.URL.Query().Get("limit"); _v != "" {
			_n, _err := strconv.ParseInt(_v, 10, 64)
			if _err != nil {
				http.Error(w, "limit"+": invalid int value: "+_err.Error(), http.StatusBadRequest)
				return
			}
			req.Limit = int(_n)
		}
		if err := req.Validate(); err != nil {
			server.WriteValidationError(w, r, err)
			return
		}
		l := logic.NewListOrdersLogic(r.Context(), svcCtx)
		resp, err := l.ListOrders(&req)
		if err != nil {
			writeError(w, err)
			return
		}
		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)
	}
}