# example/internal/handler/project-service/add-member-handler.go
// Code generated by craftgo. DO NOT EDIT.

package projectservice

import (
	"encoding/json"
	"github.com/dropship-dev/craftgo/pkg/server"
	"net/http"

	logic "github.com/dropship-dev/craftgo/example/internal/logic/project-service"
	types "github.com/dropship-dev/craftgo/example/internal/types/projects"
	"github.com/dropship-dev/craftgo/example/svccontext"
)

// AddMemberHandler returns the http.HandlerFunc for the
// POST AddMember endpoint.
func AddMemberHandler(svcCtx *svccontext.ServiceContext) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req types.AddMemberReq
		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
			http.Error(w, err.Error(), http.StatusBadRequest)
			return
		}
		req.ProjectID = r.PathValue("projectId")
		if err := req.Validate(); err != nil {
			server.WriteValidationError(w, r, err)
			return
		}
		l := logic.NewAddMemberLogic(r.Context(), svcCtx)
		resp, err := l.AddMember(&req)
		if err != nil {
			writeError(w, err)
			return
		}
		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)
	}
}