# example/internal/handler/task-service/add-comment-handler.go
// Code generated by craftgo. DO NOT EDIT.

package taskservice

import (
	"encoding/json"
	"github.com/dropship-dev/craftgo/pkg/server"
	"net/http"

	logic "github.com/dropship-dev/craftgo/example/internal/logic/task-service"
	types "github.com/dropship-dev/craftgo/example/internal/types/tasks"
	"github.com/dropship-dev/craftgo/example/svccontext"
)

// AddCommentHandler returns the http.HandlerFunc for the
// POST AddComment endpoint.
func AddCommentHandler(svcCtx *svccontext.ServiceContext) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req types.AddCommentReq
		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
			http.Error(w, err.Error(), http.StatusBadRequest)
			return
		}
		req.TaskID = r.PathValue("taskId")
		if err := req.Validate(); err != nil {
			server.WriteValidationError(w, r, err)
			return
		}
		l := logic.NewAddCommentLogic(r.Context(), svcCtx)
		resp, err := l.AddComment(&req)
		if err != nil {
			writeError(w, err)
			return
		}
		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)
	}
}