# example/internal/handler/task-service/list-tasks-handler.go
// Code generated by craftgo. DO NOT EDIT.

package taskservice

import (
	"encoding/json"
	"github.com/dropship-dev/craftgo/pkg/server"
	"net/http"
	"strconv"

	logic "github.com/dropship-dev/craftgo/example/internal/logic/task-service"
	types "github.com/dropship-dev/craftgo/example/internal/types/tasks"
	"github.com/dropship-dev/craftgo/example/svccontext"
)

// ListTasksHandler returns the http.HandlerFunc for the
// GET ListTasks endpoint.
func ListTasksHandler(svcCtx *svccontext.ServiceContext) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req types.ListTasksReq
		if _v := r.URL.Query().Get("cursor"); _v != "" {
			req.Cursor = &_v
		}
		if _v := r.URL.Query().Get("limit"); _v != "" {
			_n, _err := strconv.ParseInt(_v, 10, 64)
			if _err != nil {
				http.Error(w, "limit"+": invalid int value: "+_err.Error(), http.StatusBadRequest)
				return
			}
			req.Limit = int(_n)
		}
		if _v := r.URL.Query().Get("projectId"); _v != "" {
			req.ProjectID = &_v
		}
		if err := req.Validate(); err != nil {
			server.WriteValidationError(w, r, err)
			return
		}
		l := logic.NewListTasksLogic(r.Context(), svcCtx)
		resp, err := l.ListTasks(&req)
		if err != nil {
			writeError(w, err)
			return
		}
		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)
	}
}