# example/internal/handler/user-service/add-contact-handler.go
// Code generated by craftgo. DO NOT EDIT.

package userservice

import (
	"encoding/json"
	"github.com/dropship-dev/craftgo/pkg/server"
	"net/http"

	logic "github.com/dropship-dev/craftgo/example/internal/logic/user-service"
	types "github.com/dropship-dev/craftgo/example/internal/types/users"
	"github.com/dropship-dev/craftgo/example/svccontext"
)

// AddContactHandler returns the http.HandlerFunc for the
// POST AddContact endpoint.
func AddContactHandler(svcCtx *svccontext.ServiceContext) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req types.AddContactReq
		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
			http.Error(w, err.Error(), http.StatusBadRequest)
			return
		}
		req.UserID = r.PathValue("userId")
		if err := req.Validate(); err != nil {
			server.WriteValidationError(w, r, err)
			return
		}
		l := logic.NewAddContactLogic(r.Context(), svcCtx)
		resp, err := l.AddContact(&req)
		if err != nil {
			writeError(w, err)
			return
		}
		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)
	}
}