# example/internal/handler/user-service/list-users-handler.go
// Code generated by craftgo. DO NOT EDIT.

package userservice

import (
	"encoding/json"
	"github.com/dropship-dev/craftgo/pkg/server"
	"net/http"
	"strconv"

	logic "github.com/dropship-dev/craftgo/example/internal/logic/user-service"
	types "github.com/dropship-dev/craftgo/example/internal/types/users"
	"github.com/dropship-dev/craftgo/example/svccontext"
)

// ListUsersHandler returns the http.HandlerFunc for the
// GET ListUsers endpoint.
func ListUsersHandler(svcCtx *svccontext.ServiceContext) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req types.ListUsersReq
		if _v := r.URL.Query().Get("cursor"); _v != "" {
			req.Cursor = &_v
		}
		if _v := r.URL.Query().Get("limit"); _v != "" {
			_n, _err := strconv.ParseInt(_v, 10, 64)
			if _err != nil {
				http.Error(w, "limit"+": invalid int value: "+_err.Error(), http.StatusBadRequest)
				return
			}
			req.Limit = int(_n)
		}
		if err := req.Validate(); err != nil {
			server.WriteValidationError(w, r, err)
			return
		}
		l := logic.NewListUsersLogic(r.Context(), svcCtx)
		resp, err := l.ListUsers(&req)
		if err != nil {
			writeError(w, err)
			return
		}
		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)
	}
}